+++ NEWS.md
@@ -1,3 +1,7 @@
+# PAnEn 4.2.6
+
+- Fixed a bug when reading doubles with no decimal from an NetCDF file into R. The small error will be removed when reading.
+
 # PAnEn 4.2.5
 
 - Added `RAnEn::readNc2`, `RAnEn::readNcAttrs`, `RAnEn::mergeList`